+++ snip.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 PWD=$(cd $(dirname $0) && pwd)
-ln -s ${PWD}/haskell.snip ~/.vim/snippets/.
+[ -f "${PWD}/haskell.snip" ] && ln -s ${PWD}/haskell.snip ~/.vim/snippets/.
